pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncmy>=0.2.10",
    "fastapi>=0.118.0",
    "passlib[bcrypt]",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.9",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
    "click>=8.1.7",
    "aiomysql>=0.2.0",
    "bcrypt>=3.2.0",
]
